Moved initalization function back into driver
